Fix CMake (#251)
The CMake instructions in the readme didn't quite work when I tried them on macOS today.

This patch includes a few changes to `CMakeLists.txt` and the `README.md` instructions to reflect the latest steps to build.

Also add a CI job to validate that this keeps working moving forward.