build: add the ability to run the test suite from CMake
Expand the CMake based build to support running the test suite.  This
allows testing on Windows as well.  Although the test suite does not
fully pass on Windows yet, this brings us closer.

With this on Windows, the following is the result:

```
  Test Suite 'All tests' failed at 2020-05-14 12:11:08.871
           Executed 120 tests, with 9 failures (0 unexpected) in 0.505 (0.505) seconds
```

The failures in the test suite correspond to the conversion of floating
point values in the infinity and NaN cases (either converting to
`-infe+0`, `infe+0`, or `nane+0` instead of `-inf`, `inf`, `nan`, or
converting to `NaN` instead of `nan`).