Add `sequenceStyle` and `mappingStyle` to `Emitter.Options` (#340)
## Summary:

This adds `sequenceStyle` and `mappingStyle` options to `Emitter.Options`, which allows those using `YAMLEncoder` to specify whether "sequences" and "mappings" use `block` style (the default style) or `flow` style.

Also, CHANGELOG.md was updated and Podspec version was bumped to 4.0.7.

No unit test coverage was added, as there isn't direct unit test coverage for `Emitter.Options`, however there are already tests around mapping and sequence styles in the `EmitterTests` class, so this should be sufficient.

_**Note:** "sequences" are arrays / lists, "mappings" are dictionaries_

## Strategy:

`sequenceStyle` and `mappingStyle` were added to `Emitter.Options`, and changes were cascaded down into `Emitter`, `Encoder`, and `Node` where necessary. The existing field `sortKeys` in `Emitter.Options` was used as a guide to see where both `sequenceStyle` and `mappingStyle` should be added in other places in the code. Otherwise, minimal code changes were made to gain the desired feature.

## Examples:

Given an example YAML file here:

```
structure:
  edges:
    e1:
    - n1
    - n2
    e2:
    - n1
    - n3
    e3:
    - n2
    - n3
  nodes:
  - n1
  - n2
  - n3
```

### `sequenceStyle`

Setting `sequenceStyle` to `.flow` now generates:

```
structure:
  edges:
    e1: [n1, n2]
    e2: [n1, n3]
    e3: [n2, n3]
  nodes: [n1, n2, n3]
```

Which can be more visually pleasing, readable, and compact in certain scenarios such as this one.

### `mappingStyle`

Setting `mappingStyle` to `.flow` now generates:

```
{structure: {edges: {e1: [n1, n2], e2: [n1, n3], e3: [n2, n3]}, nodes: [n1, n2, n3]}}
```

Which may not be as visually pleasing with larger YAML files, but still could be useful in certain scenarios, and having the option to configure `mappingStyle` does "match parity" with being able to configure `sequenceStyle`.

Also you might notice, setting `mappingStyle` to `.flow` also forces `sequenceStyle` to be `.flow`, but this is not an error as for "flow" style to be used correctly with mappings at the top level, it must be enforced recursively through each child. (Implying, child mappings and child sequences must also be "flow" style.)