Yams: make `CYaml` private to Yams (#343)
This changes how `CYaml` is built and used within Yams.  As there are no
interfaces from `CYaml` being directly exposed, we can statically link
the library.  However, this still would cause a problem as the `CYaml`
import would be serialized into the module requiring that the module is
available when building anything which consumes `Yams`.  To avoid that,
change the `import CYaml` instances to `@_implementationOnly`.  This
allows for `CYaml` to not be required at the consumer site.

While in the area of the CMake build system, clean up some of the build
system to use language specific options.  This is required to avoid the
de-duplication of the options across languages and enables linking
`CYaml` statically into the library.  Note that on Windows static
linking of *Swift* libraries is not yet properly supported and this does
not enable that nor use that - it is statically linking *C* content.

This now allows building Yams dynamically as a single library target.
On Windows, this is a negligible space savings of ~16KiB.